@@ -1,5 +1,5 @@
 import { Component, ElementRef, Input, OnInit, ViewChild } from '@angular/core';
-import { takeUntil, tap } from 'rxjs/operators';
+import { filter, takeUntil, tap } from 'rxjs/operators';
 import { FormControl, FormGroup } from '@angular/forms';
 import { MatAutocompleteSelectedEvent } from '@angular/material';
 
@@ -35,7 +35,10 @@ export class OperatorsAutocompleteComponent extends DestroyObservable implements
     this.loadOperators();
     this.operatorCtrl.valueChanges
       .pipe(takeUntil(this.destroy$))
-      .pipe(tap((literal) => this.filterOperators(literal)))
+      .pipe(
+        filter((literal) => !!literal),
+        tap((literal: string) => this.filterOperators(literal)),
+      )
       .subscribe();
   }
 
@@ -44,7 +47,8 @@ export class OperatorsAutocompleteComponent extends DestroyObservable implements
   }
 
   /**
-   * todo: refactor when search is made server side
+   * todo: r
+   * efactor when search is made server side
    */
   getOperatorLabel(operatorId: string): string {
     return this.operators.find((operator) => operator._id === operatorId).nom_commercial;
@@ -83,7 +87,8 @@ export class OperatorsAutocompleteComponent extends DestroyObservable implements
     const selectedOperatorIds = this.operatorIdsControl.value || [];
     this.filteredOperators = this.operators.filter(
       (operator) =>
-        selectedOperatorIds.indexOf(operator._id) === -1 && operator.nom_commercial.toLowerCase().includes(literal),
+        selectedOperatorIds.indexOf(operator._id) === -1 &&
+        operator.nom_commercial.toLowerCase().includes(literal.toLowerCase()),
     );
   }
 }
