@@ -1,14 +1,14 @@
-import { Address, Company, Territory } from '../../../../src/app/core/entities/territory/territory';
+import { Contact } from '~/core/entities/shared/contact';
+import { JsonRPCResponse } from '~/core/entities/api/jsonRPCResponse';
 
-import { JsonRPCResponse } from '../../../../src/app/core/entities/api/jsonRPCResponse';
+import { Address, Company, Contacts, Territory } from '../../../../src/app/core/entities/territory/territory';
 
-export const territoryStub: Territory = {
+export const territoryStub: Territory = new Territory({
   _id: '5c66d89760e6ee004a6cab2f',
   name: 'AOM 1',
-  acronym: 'Aom 1 acronym ',
   shortname: 'AOM 1 shortname',
   company: new Company({
-    siren: '123456789',
+    siret: '12345678915654654',
     naf_entreprise: '1234A',
   }),
   address: new Address({
@@ -17,25 +17,25 @@ export const territoryStub: Territory = {
     city: 'Lyon',
     country: 'France',
   }),
-  contacts: {
-    gdpr_controller: {
+  contacts: new Contacts({
+    gdpr_controller: new Contact({
       firstname: 'Raymond',
       lastname: 'Breton',
       email: 'raymond.breton@aom.com',
-    },
-    gdpr_dpo: {
+    }),
+    gdpr_dpo: new Contact({
       firstname: 'Michelle',
       lastname: 'Planche',
       email: 'michelle.planche@aom.com',
-    },
-    technical: {
+    }),
+    technical: new Contact({
       firstname: 'Albert',
       lastname: 'Marcoeur',
       email: 'albert.marcoeur@aom.com',
       phone: '0673826458',
-    },
-  },
-};
+    }),
+  }),
+});
 
 export function stubTerritoryFind() {
   cy.route({
